0000044: Lieferant: gamesplanet * Added module for partner gamesplanet * Added affiliate tools to affiliate modul for offering some standard functionality for all partner requests
git-svn-id: https://svn.gate4games.com/svn/Gate4Games/trunk@332 22d369d5-c360-4d0a-819e-37d15cbb1d4a

+++ htdocs/modules/lv/lvAffiliate/application/models/lvaffiliate_import.php
@@ -308,18 +308,20 @@ class lvaffiliate_import extends oxBase {
             }
 
             if ( $this->_sLvCurrentParentId ) {
-                foreach ( $this->_aLvCurrentArticleData[$sDataFieldName] as $sTargetCategoryId ) {
-                    $blAssignmentExists = $this->_lvCheckCategoryAssignmentExists( $sTargetCategoryId );
-                    
-                    if ( $blAssignmentExists === false ) {
-                        $sNewId = $oUtilsObject->generateUId();
-                        
-                        $sQuery = "INSERT INTO oxobject2category ( OXID, OXOBJECTID, OXCATNID ) VALUES ( '".$sNewId."', '".$this->_sLvCurrentParentId."', '".$sTargetCategoryId."' )";
-                        $oDb->Execute( $sQuery );
-                    }
-                    else if ( $blCheckForRemoval === true ) {
-                        $sQuery = "DELETE FROM oxobject2category WHERE OXOBJECTID='".$this->_sLvCurrentParentId."' AND OXCATNID='".$sTargetCategoryId."' LIMIT 1";
-                        $oDb->Execute( $sQuery );
+                if ( isset( $this->_aLvCurrentArticleData[$sDataFieldName] ) && is_array( $this->_aLvCurrentArticleData[$sDataFieldName] ) ) {
+                    foreach ( $this->_aLvCurrentArticleData[$sDataFieldName] as $sTargetCategoryId ) {
+                        $blAssignmentExists = $this->_lvCheckCategoryAssignmentExists( $sTargetCategoryId );
+
+                        if ( $blAssignmentExists === false ) {
+                            $sNewId = $oUtilsObject->generateUId();
+
+                            $sQuery = "INSERT INTO oxobject2category ( OXID, OXOBJECTID, OXCATNID ) VALUES ( '".$sNewId."', '".$this->_sLvCurrentParentId."', '".$sTargetCategoryId."' )";
+                            $oDb->Execute( $sQuery );
+                        }
+                        else if ( $blCheckForRemoval === true ) {
+                            $sQuery = "DELETE FROM oxobject2category WHERE OXOBJECTID='".$this->_sLvCurrentParentId."' AND OXCATNID='".$sTargetCategoryId."' LIMIT 1";
+                            $oDb->Execute( $sQuery );
+                        }
                     }
                 }
             }
